Adding showalter to thermo.py and test_thermo.py, removing solor file
