+++ src/metpy/calc/thermo.py
@@ -3264,3 +3264,66 @@ def gradient_richardson_number(height, potential_temperature, u, v, vertical_dim
     dvdz = first_derivative(v, x=height, axis=vertical_dim)
 
     return (mpconsts.g / potential_temperature) * (dthetadz / (dudz ** 2 + dvdz ** 2))
+
+@exporter.export
+@preprocess_and_wrap()
+@check_units('[pressure]', '[temperature]', '[temperature]')
+def showalter_index(pressure, temperature, dewpt): 
+    """Calculate Showalter Index from pressure temperature and 850 hPa lcl
+    
+    Showalter Index derived from [Galway1956]_:
+    SI = T500 - Tp500
+    
+    where:
+    T500 is the measured temperature at 500 hPa
+    Tp500 is the temperature of the lifted parcel at 500 hPa
+    
+   Parameters
+   ----------
+        
+        pressure : `pint.Quantity`
+            Atmospheric pressure level(s) of interest, in order from highest to
+        lowest pressure
+        
+        temperature : `pint.Quantity`
+            Parcel temperature for corresponding pressure 
+        
+        dewpt (:class: `pint.Quantity`):
+            Parcel dew point temperatures for corresponding pressure
+        
+
+    Returns
+    -------
+    `pint.Quantity`
+        Showalter index in delta degrees celsius
+        
+    """
+
+    # find the measured temperature and dew point temperature at 850 hPa.
+    idx850 = np.where(pressure == 850 * units.hPa)
+    T850 = temperature[idx850]
+    Td850 = dewpt[idx850]
+    
+    # find the parcel profile temperature at 500 hPa.
+    idx500 = np.where(pressure == 500 * units.hPa)
+    Tp500 = temperature[idx500]
+    
+    # Calculate lcl at the 850 hPa level
+    lcl_calc = lcl(850 * units.hPa, T850[0], Td850[0])
+    lcl_calc = lcl_calc[0]
+    
+    # Define start and end heights for dry and moist lapse rate calculations
+    p_strt = 1000 * units.hPa
+    p_end = 500 * units.hPa
+    
+    # Calculate parcel temp when raised dry adiabatically from surface to lcl
+
+    dl = dry_lapse(lcl_calc, tc[0], p[0])
+    dl = (dl.magnitude - 273.15) * units.degC  # Change units to C
+
+    # Calculate parcel temp when raised moist adiabatically from lcl to 500mb
+    ml = moist_lapse(p_end, dl, lcl_calc)
+    
+    # Calculate the Showalter index
+    shox = Tp500 - ml
+    return shox
