+++ tests/calc/test_thermo.py
@@ -26,7 +26,7 @@ from metpy.calc import (brunt_vaisala_frequency, brunt_vaisala_frequency_squared
                         relative_humidity_wet_psychrometric,
                         saturation_equivalent_potential_temperature,
                         saturation_mixing_ratio,
-                        saturation_vapor_pressure,
+                        saturation_vapor_pressure, showalter_index,
                         specific_humidity_from_dewpoint,
                         specific_humidity_from_mixing_ratio, static_stability,
                         surface_based_cape_cin, temperature_from_potential_temperature,
@@ -1843,3 +1843,14 @@ def test_gradient_richardson_number_with_xarray():
     assert isinstance(result, xr.DataArray)
     xr.testing.assert_identical(result['height'], data['Ri_g']['height'])
     assert_array_almost_equal(result.data.m_as(''), data['Ri_g'].data)
+
+def test_showalter_index():
+    p_upper = np.arange(1000, 200, -50) * units.hPa 
+    p_lower = np.arange(175, 0, -25) * units.hPa 
+    p = np.append(p_upper, p_lower,)
+    tc = np.linspace(30, -30, 25) * units.degC 
+    tdc = np.linspace(10, -30, 25) * units.degC
+    
+    result = showalter_index(p, tc, tdc)
+    expected = 23.73036498600014 * units.degC
+    assert_array_almost_equal(result, expected, 4)
